Added MonsterList Class Compenent
The MonsterList Class Component is put in the RoomComponent class and is used to show the monster data through monster objects as json data. it reads the input and puts it on display on the page and allows the user to change the inputs if nessecary. Not able to add new monster quite yet